moved logic back into currencyRatesFileGen.js and added gitignore
