+++ .gitignore
@@ -1,0 +1,79 @@
+# Built Files
+node_modules/
+build
+
+# Test Files
+test/app
+gpt.html
+gpt-each-bidder3.html
+
+# Selenium files
+bin
+selenium*.log
+
+# Dev File
+integrationExamples/gpt/gpt.html
+integrationExamples/gpt/*-test.html
+integrationExamples/implementations/
+src/adapters/analytics/libraries
+
+# Coverage reports
+build/coverage/
+
+# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion
+
+*.iml
+
+## Directory-based project format:
+.idea/
+# if you remove the above rule, at least ignore the following:
+
+# User-specific stuff:
+# .idea/workspace.xml
+# .idea/tasks.xml
+# .idea/dictionaries
+
+# Sensitive or high-churn files:
+# .idea/dataSources.ids
+# .idea/dataSources.xml
+# .idea/sqlDataSources.xml
+# .idea/dynamic.xml
+# .idea/uiDesigner.xml
+
+# Gradle:
+# .idea/gradle.xml
+# .idea/libraries
+
+# Mongo Explorer plugin:
+# .idea/mongoSettings.xml
+
+## File-based project format:
+*.ipr
+*.iws
+
+## Plugin-specific files:
+
+# IntelliJ
+/out/
+
+# mpeltonen/sbt-idea plugin
+.idea_modules/
+
+# JIRA plugin
+atlassian-ide-plugin.xml
+
+# Crashlytics plugin (for Android Studio and IntelliJ)
+com_crashlytics_export_strings.xml
+crashlytics.properties
+crashlytics-build.properties
+*.swp
+*.swo
+
+#zip files (for releases)
+*.zip
+
+# TypeScript typings
+typings/
+
+# MacOS system files
+.DS_Store
