+++ src/currencyRatesFileGen.js
@@ -1,21 +1,23 @@
 /**
- * @fileOverview Loads currency rates from the European Central Bank and creates and uploads a JSON representation to S3.
- * @version 1.1
- */
+ ** @fileOverview Loads currency rates from the European Central Bank and uploads a JSON representation to github.
+ ** @version 2.0
+ **/
+'use strict';
 const https = require('https');
 const aws = require('aws-sdk');
+const fs = require('fs')
+const path = require('path')
+const process = require('process')
+const { spawnSync } = require('child_process')
+
+const { GITHUB_TOKEN, GITHUB_USERNAME, GITHUB_EMAIL } = process.env
+// leaving this without https:// in order to reuse it when adding the remote
+const gitRepositoryURL = 'github.com/prebid/currency-file.git'
+const repositoryName = 'currency-file'
 
 const fromCurrencies = ['USD', 'GBP'];
-
 // when to expire for HTTP "Expires:" header (seconds)
 const expires = 24 * 3600 + 5;
-
-/**
- * @returns {boolean} env variable value if set or default
- */
-function getDebug() {
-    return (typeof process.env.DEBUG !== 'undefined') ? (process.env.DEBUG === '1') : true;
-}
 
 /**
  *  env variable value if set or default
@@ -34,86 +36,162 @@ function getFilename() {
 }
 
 /**
- * The function AWS Lambda calls to start execution of your Lambda function.
- * You identify the handler when you create your Lambda function: IE
- * 'Handler':'currencyRatesFilesGen.handler'
- * @param event - AWS Lambda uses this parameter to pass in event data to the handler.
- * @param context - the context parameter contains functions to access runtime information
- */
-exports.handler = function(event, context) {
-    /** @type {Array.<Object>} - loaded and parsed json objects for currency */
-    const results = [];
-    console.log('fromCurrencies', fromCurrencies);
+* @returns {boolean} env variable value if set or default
+*/
+function getDebug() {
+    return (typeof process.env.DEBUG !== 'undefined') ? (process.env.DEBUG === '1') : true;
+}
 
-    for (let fromCurrency of fromCurrencies) {
-        /** @type {string|undefined} */
-        const currencyUrl = constructCurrencyUrl(fromCurrency);
-        if (!currencyUrl) {
-            logError('Error: malformed currencyUrl', currencyUrl);
-            // error, exit
-            return;
-        }
-
-        // load currency json file from currency url
-        requestCurrencyFile(currencyUrl, (json) => {
-	    // verify response data, json.rates should be an object with at least 20 keys (currencies)
-	    if (json !== null && typeof json === 'object') {
-	      if (json.base && json.date && typeof json.rates === 'object' && Object.keys(json.rates).length >= 20) {
-	        results.push(json);
-	       }
-	       else {
-	        logError('Error: json data failed validation:', json);
-	        return;
-	       }
-	    }
-
-            // All results loaded when results count is equal to fromCurrencies count
-            if (results.length === fromCurrencies.length) {
-                const expiration = getExpiration(expires);
-                if (!expiration) {
-                    logError('Error: malformed expiration date:', expiration);
-                    // error, exit
-                    return;
-                }
-
-                const docParams = createDocumentParams(getBucket(), getFilename(), createDocument(results), expiration);
-                if (!docParams) {
-                    logError('Error: malformed docParams:', docParams);
-                    // error, exit
-                    return;
-                }
-
-                // upload json to S3 bucket at key
-                // context.done() is called on complete/error
-                uploadDocumentToS3(docParams, context);
-            } else {
-                logError('Error: did not receive responses for all fromCurrencies');
-                return;
-	    }
-        });
-    }
-};
 
 /**
- * @param {string} bucket
- * @param {string} filename
- * @param {Object} documentObj
- * @param {Date} expires - when to expire for HTTP "Expires:"
- * @return {S3UploadParams|undefined}
+ * runs a command line function
+ * 
+ * @param {any} commandString 
+ * @param {any} options 
  */
-function createDocumentParams(bucket, filename, documentObj, expires) {
-    if (typeof bucket === 'string' && bucket !== '' &&
-        typeof filename === 'string' && filename !== '' &&
-        documentObj !== null && typeof documentObj === 'object' && !Array.isArray(documentObj) && Object.keys(documentObj).length > 0 &&
-        documentObj !== null && typeof expires === 'object' && expires instanceof Date) {
-        return {
-            Bucket: bucket,
-            Key: filename,
-            Body: JSON.stringify(documentObj),
-            Expires: expires
-        }
+function runCommand(commandString, options) {
+    const [command, ...args] = commandString.match(/(".*?")|(\S+)/g)
+    const cmd = spawnSync(command, args, options)
+    const errorString = cmd.stderr.toString()
+    if (errorString) {
+        console.log('throwing error', errorString);
+        throw new Error(
+            `Git command failed
+      ${commandString}
+      ${errorString}`
+        )
     }
 }
+
+// gets the current git version and increments the third level
+function incGitTag() {
+    var result = spawnSync('git', ['describe', '--abbrev=0', '--tags'], { stdio: 'pipe' });
+    var errorString = result.stderr.toString()
+    if (errorString) {
+        logError("incGitTag error: " + errorString);
+        return (-1);
+    }
+    var output = result.stdout.toString();
+    if (output && output != "") {
+        // console.log("incGitTag output: " + output);
+        var versionArray;
+        versionArray = output.split(".");
+        if (versionArray.length != 3) {
+            logError("invalid version: " + version);
+            return (-1);
+        }
+        var version3 = parseInt(versionArray[2]);
+        if (isNaN(version3)) {
+            logError("invalid version: " + version);
+            return (-1);
+        }
+        console.log('NEW VERSION');
+
+        var newVersion = versionArray[0] + "." + versionArray[1] + "." + (version3 + 1);
+        console.log(newVersion);
+        // write tag
+        result = spawnSync('git', ['tag', newVersion]);
+        var errorString = result.stderr.toString()
+        if (errorString) {
+            logError("git tag error: " + errorString);
+            return (-1);
+        }
+        console.log("version updated to " + newVersion);
+    } else {
+        logError("git describe output was invalid");
+        return (-1);
+    }
+}
+
+module.exports.downloadPublish = async function (event, context, callback) {
+    process.env['PATH'] = process.env['PATH'] + ':' + process.env['LAMBDA_TASK_ROOT']
+
+    // install git binary
+    await require('lambda-git')()
+
+    // first try to get the data from the source
+
+    /** @type {Array.<Object>} - loaded and parsed json objects for currency */
+    //console.log('fromCurrencies', fromCurrencies);
+    const urls = [];
+    fromCurrencies.forEach((fromCurrency)=> urls.push(constructCurrencyUrl(fromCurrency)));
+    let responses;
+    try{
+        responses = await fetchMultipleUrlsAsPromise(urls);
+    }
+    catch(e){
+        context.fail(e);
+        return;
+    }
+    const expiration = getExpiration(expires);
+    if (!expiration) {
+        logError(`Error: malformed expiration date: ${expiration}`, context);
+        return;
+    }
+
+    const newDocument = createDocument(responses);
+
+    // for a transition period we send to github and then write to S3
+    if (pushToGithub(newDocument) < 0) {
+        logError("failure pushing to github", context);
+        return;
+    }
+
+    const docParams = createDocumentParams(getBucket(), getFilename(), newDocument, expiration);
+    if (!docParams) {
+        const error = 'Error: malformed docParams:';
+        logError(error, context);
+        return;
+    }
+    // upload json to S3 bucket at key
+    // context.done() is called on complete/error
+    uploadDocumentToS3(docParams, context);
+
+    return "success";
+}
+
+// Based on https://gist.github.com/Loopiezlol/e00c35b0166b4eae891ec6b8d610f83c
+// TODO clean up and add tests
+function pushToGithub(newDocument) {
+
+    // now push the file up to git
+    // change the cwd to /tmp
+    process.chdir('/tmp')
+    // clone the repository and set it as the cwd
+    runCommand(`git clone --quiet https://${gitRepositoryURL}`);
+    process.chdir(path.join(process.cwd(), repositoryName))
+    runCommand(`git pull`);
+    // update local file
+    fs.writeFileSync('latest.json', JSON.stringify(newDocument));
+
+    // update local git config with email and username (required)
+    runCommand(`git config --local user.email ${GITHUB_EMAIL}`)
+    runCommand(`git config --local user.name ${GITHUB_USERNAME}`)
+    // stage local files
+    runCommand('git add .')
+    // commit changes
+    runCommand('git commit -m "commit by :robot:"')
+    if (incGitTag() < 0) {
+        logError("incGitTag failed");
+        return (-1);
+    }
+    // replace the remote with an authenticated one
+    runCommand('git remote rm origin')
+    runCommand(
+        `git remote add origin https://${GITHUB_USERNAME}:${GITHUB_TOKEN}@${gitRepositoryURL}`
+    )
+    // push changes to remote
+    runCommand('git push --porcelain --set-upstream origin master')
+    runCommand('git push --porcelain origin --tags')
+    return (0);
+}
+
+// TODO - clear jsdelivr cache
+// Purge cache
+// jsDelivr has an easy to use API to purge files from the cache and force the files to update. This is useful when you release a new version and want to force the update of all version aliased users.
+//
+// To avoid abuse, access to purge is given after an email request (for now - dak@prospectone.io).
+
 
 /**
  * @param {number} expires - when to expire for HTTP "Expires:" header (seconds)
@@ -141,6 +219,7 @@ function constructCurrencyUrl(fromCurrency) {
  * @param {string} url
  * @param {function} fileEndCallback
  * @returns {http.ClientRequest}
+ * TODO: add promise style rejections. 
  */
 function requestCurrencyFile(url, fileEndCallback) {
     log('requesting: ' + url);
@@ -178,44 +257,6 @@ function requestCurrencyFile(url, fileEndCallback) {
 }
 
 /**
- * @typedef {Object} S3UploadParams
- * @property {string} Bucket - Name of the bucket to which the PUT operation was initiated.
- * @property {string} Key - Object key for which the PUT operation was initiated.
- * @property {Buffer|Array|Blob|String|ReadableStream} Body - Object data
- * @property {Date} Expires - The date and time at which the object is no longer cacheable.
- */
-
-/**
- * @param {S3UploadParams} params
- * @param context
- */
-function uploadDocumentToS3(params, context) {
-    if (params === null || typeof params !== 'object') {
-        logError('Error: invalid params argument passed to uploadDocumentToS3', params);
-        return;
-    }
-    if (context === null || typeof context !== 'object' || !context.hasOwnProperty('done') || typeof context.done !== 'function') {
-        logError('Error: invalid context argument passed to uploadDocumentToS3', context);
-        return;
-    }
-
-    log('rates s3 upload to: ' + params.Bucket + ' ' + params.Key);
-
-    // Upload assembled currency json to S3 bucket
-    const s3 = new aws.S3();
-    // upload callback executes context.done on complete or error
-    return s3.upload(params, (e, data) => {
-        if (e) {
-            logError(e.lineNumber, e.message);
-            context.done(null, JSON.stringify({filename: params.Key, error: data}));
-        } else {
-            log('rates pushed to s3: ' + params.Bucket + ' ' + params.Key);
-            context.done(null, JSON.stringify({filename: params.Key}));
-        }
-    });
-}
-
-/**
  * @param {Array.<Object>} results
  * @returns {{dataAsOf: string, conversions: {}}}
  */
@@ -238,27 +279,112 @@ function log(line) {
 }
 
 /**
- * @param {*} line
- * @param {*} error
+ * @param {String} error
+ * @param {Object} Context
  */
-function logError(line, error) {
-    console.error(line, error);
+function logError(error, context) {
+    console.error(error);
+    if(context && typeof context === 'object' && typeof context.fail === 'function' ) {
+        context.fail(error);
+    }
 }
 
+/**
+ * @typedef {Object} S3UploadParams
+ * @property {string} Bucket - Name of the bucket to which the PUT operation was initiated.
+ * @property {string} Key - Object key for which the PUT operation was initiated.
+ * @property {Buffer|Array|Blob|String|ReadableStream} Body - Object data
+ * @property {Date} Expires - The date and time at which the object is no longer cacheable.
+ */
+/**
+ * @param {S3UploadParams} params
+ * @param context
+ */
+function uploadDocumentToS3(params, context) {
+    if (params === null || typeof params !== 'object') {
+        logError(`Error: invalid params argument passed to uploadDocumentToS3: ${params}`, context);
+        return;
+    }
+    if (context === null || typeof context !== 'object' || !context.hasOwnProperty('done') || typeof context.done !== 'function') {
+        logError('Error: invalid context argument passed to uploadDocumentToS3', context);
+        return;
+    }
+
+    log('rates s3 upload to: ' + params.Bucket + ' ' + params.Key);
+
+    // Upload assembled currency json to S3 bucket
+    const s3 = new aws.S3();
+    // upload callback executes context.done on complete or error
+    return s3.upload(params, (e, data) => {
+        if (e) {
+            logError(e.message, context);
+            context.done(null, JSON.stringify({ filename: params.Key, error: data }));
+        } else {
+            log('rates pushed to s3: ' + params.Bucket + ' ' + params.Key);
+            context.done(null, JSON.stringify({ filename: params.Key }));
+        }
+    });
+}
+
+/**
+ * @param {string} bucket
+ * @param {string} filename
+ * @param {Object} documentObj
+ * @param {Date} expires - when to expire for HTTP "Expires:"
+ * @return {S3UploadParams|undefined}
+ */
+function createDocumentParams(bucket, filename, documentObj, expires) {
+    if (typeof bucket === 'string' && bucket !== '' &&
+        typeof filename === 'string' && filename !== '' &&
+        documentObj !== null && typeof documentObj === 'object' && !Array.isArray(documentObj) && Object.keys(documentObj).length > 0 &&
+        documentObj !== null && typeof expires === 'object' && expires instanceof Date) {
+        return {
+            Bucket: bucket,
+            Key: filename,
+            Body: JSON.stringify(documentObj),
+            Expires: expires
+        }
+    }
+}
+
+async function fetchMultipleUrlsAsPromise(fetchUrls){
+    const responses = [];
+    for (let url of fetchUrls) {
+        const promise = new Promise((resolve, reject) => {
+            requestCurrencyFile(url, resolve);
+        });
+        const data = await promise;
+        validateCurrencyData(data);
+        responses.push(data);
+    }
+    return responses;
+}
+
+function validateCurrencyData(json) {
+    if(!json || typeof json !== 'object') {
+        throw new Error(`Error: json data failed validation`);
+    }
+    if (!json.base || !json.date || typeof json.rates !== 'object' || Object.keys(json.rates).length < 20) {
+        throw new Error(`Error: json data failed validation: ${json}`);
+    }
+}
 
 /**
  * Export internal functions for testing
  */
 exports.spec = {
     getDebug,
-    getFilename,
-    getBucket,
     log,
     logError,
     constructCurrencyUrl,
     requestCurrencyFile,
-    uploadDocumentToS3,
-    createDocumentParams,
+    pushToGithub,
     getExpiration,
-    createDocument
+    createDocument,
+    runCommand,
+    incGitTag,
+    uploadDocumentToS3,
+    getFilename,
+    getBucket,
+    createDocumentParams
 };
